Revert to using attr for fill
I've tested this on Chrome and Firefox. In both cases, fill color is set
by one of the following rules, in order of highest to lowest priority:

1. Fill is defined in the style attribute
2. Fill is an attribute itself

When we switched fill to a style it breaks the application of CSS rules.
By leaving it as an attribute, we get a default of a white background,
apply any CSS rules, and then override those rules only when 'fill' has
been set on the node itself.