Fix #66: dagre-d3 changes attributes in the input graph
The bug fix itself is in lib/Renderer.js where we now more aggresively
copy the input graph. This is still not a deep copy and it is
potentially less efficient than necessary.

The major of this change is adding support for running mocha tests in
the browser and from the command line.