Remove safari karma launcher
Occasionally, with the Safari launcher, Safari gets stuck in a state
where it cannot be stopped, doesn't appear as a process, etc. but is
still hanging around in the OSX app switcher. In this state it is
impossible to launch a new instance of Safari. For now I'm pulling
Safari launcher until this gets resolved.