Initial work on label justification
This change allows labels to be justified for clusters and for nodes. It
is a work in progress. The following problems exist:

1. We don't ensure that the cluster is large enough for the label. This
could be done by setting the top or bottom dummy node of the cluster to
the width of the label + padding. This must be done in dagre.

2. Similar problem for making vertical space for the label in the top or
bottom position. This was not introduced with this change, but it could
be fixed in dagre at the same time. Proposed solution: use the height +
padding at the top or bottom dummy node as appropriate. This may cause
some unrelated borders to have extra padding.

3. Misc bugs when using labelJust on nodes. E.g. position gets moved to
the opposite of justfication and the label is moved towards the bottom.