Bug fix: removing clusters should not throw
Prior to this commit we selected clusters for positioning using the
"g.cluster" selector. This would select removed clusters because these
removed clusters remain in the DOM until they've completely faded out.
However, the graph knows nothing about removed clusters, so we'd throw
an exception when trying to get the position of the cluster in
positionClusters.

We fix this by getting the selection from createClusters, similar to how
we get the selection for nodes. The createClusters selection set is
guaranteed to only include clusters that are in the graph.